Clippy, deps & stuff (#100)
* explain unsafe block in merkle-hash-bench/src/mod_ring/ruint.rs

* add missing_safety_doc lint

* clippy autofix

* disable overzealous lints

* fix some lints

* remove unused code

* lints contd.

* refactor: update member Cargo.toml files to use workspace dependencies

* feat: include workspace members as dependencies in Cargo.toml

* use workspace deps

* fmt

* downgrade bincode

* cargo check

* WIP: fix clippy lints

* fix lints

* autofix

* static assertions

* allow tranmuste

* allow range loop

* fix doctest

* fix readme doc test

* lock git deps

* tmp: disable icicle

* nuke merkle-hash-bench

+++ noir-r1cs/src/range_check.rs
@@ -231,7 +231,7 @@ fn add_naive_range_check(
         &[
             (FieldElement::one(), index_witness),
             (
-                FieldElement::from((1 << num_bits) - 1 as u32).neg(),
+                FieldElement::from((1 << num_bits) - 1_u32).neg(),
                 r1cs_compiler.witness_one(),
             ),
         ],
